refactor: extract first-use CLI modules
