+++ crates/perfgate-cli/src/main.rs
@@ -1,9 +1,13 @@
 //! perfgate CLI - entry point for all workflows.
 
+mod artifact_explain;
 mod baseline;
 mod check_guidance;
 mod cli_parsing;
+mod decision_suggest;
 mod doctor;
+mod ledger_doctor;
+mod probe_templates;
 mod repair_context;
 mod storage;
 
@@ -13,14 +17,17 @@ use storage::{
 };
 
 use anyhow::Context;
+use artifact_explain::execute_explain_action;
 use baseline::{BaselineSelector, parse_baseline_selector};
 use check_guidance::{
     FailureClass, check_command, classify_check_error, emit_check_outcome_guidance, paired_command,
-    print_check_failure_guidance, shell_path,
+    print_check_failure_guidance,
 };
 use clap::{Args, Parser, Subcommand, ValueEnum};
 use cli_parsing::*;
+use decision_suggest::execute_decision_suggest;
 use glob::glob;
+use ledger_doctor::execute_ledger_action;
 use perfgate::app as perfgate_app;
 use perfgate::domain as perfgate_domain;
 use perfgate::integrations::github::{self, CommentOptions, GitHubClient};
@@ -51,8 +58,8 @@ use perfgate_client::types::{BaselineRecord, CreateKeyRequest, KeyEntry};
 use perfgate_client::{
     AuthMethod, BaselineClient, ClientConfig, ListAuditEventsQuery, ListAuditEventsResponse,
     ListBaselinesQuery, ListDecisionsQuery, ListVerdictsQuery, PruneDecisionsRequest,
-    ResolvedServerConfig, RetryConfig, SubmitVerdictRequest, UploadBaselineRequest,
-    UploadDecisionRequest, resolve_server_config,
+    ResolvedServerConfig, SubmitVerdictRequest, UploadBaselineRequest, UploadDecisionRequest,
+    resolve_server_config,
 };
 use perfgate_domain::scaling::{
     ScalingReport, SizeMeasurement, classify_complexity, parse_complexity, render_ascii_chart,
@@ -85,10 +92,8 @@ use std::process::ExitCode;
 use std::sync::Arc;
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 
-use doctor::{
-    DoctorCheck, DoctorStatus, ensure_artifact_dir_writable, execute_calibrate, execute_doctor,
-    plural,
-};
+use doctor::{DoctorCheck, DoctorStatus, execute_calibrate, execute_doctor, plural};
+use probe_templates::execute_probe_init;
 
 const BASELINE_SERVER_NOT_CONFIGURED: &str = "baseline server is not configured; set `--baseline-server`, `PERFGATE_SERVER_URL`, or `[baseline_server].url` in `perfgate.toml`";
 const DEFAULT_FALLBACK_BASELINE_DIR: &str = "baselines";
@@ -3205,239 +3210,6 @@ fn execute_probe_action(action: ProbeAction) -> anyhow::Result<()> {
     }
 }
 
-fn execute_probe_init(args: ProbeInitArgs) -> anyhow::Result<()> {
-    let template = probe_template(args.template);
-    fs::create_dir_all(&args.out_dir)
-        .with_context(|| format!("create probe template directory {}", args.out_dir.display()))?;
-
-    write_probe_template_file(
-        &args.out_dir.join("README.md"),
-        &render_probe_template_readme(&template, &args.out_dir),
-        args.force,
-    )?;
-    write_probe_template_file(
-        &args.out_dir.join("probes-baseline.jsonl"),
-        &format!("{}\n", template.baseline_jsonl.join("\n")),
-        args.force,
-    )?;
-    write_probe_template_file(
-        &args.out_dir.join("probes-current.jsonl"),
-        &format!("{}\n", template.current_jsonl.join("\n")),
-        args.force,
-    )?;
-    write_probe_template_file(
-        &args.out_dir.join("scenario.toml"),
-        &render_probe_template_scenario(&template),
-        args.force,
-    )?;
-    write_probe_template_file(
-        &args.out_dir.join("tradeoff.toml"),
-        &render_probe_template_tradeoff(&template),
-        args.force,
-    )?;
-
-    eprintln!(
-        "Probe starter template '{}' written to {}",
-        template.slug,
-        args.out_dir.display()
-    );
-    eprintln!("Review the generated snippets before adding them to perfgate.toml.");
-    eprintln!(
-        "Next: perfgate ingest probes --file {}/probes-current.jsonl --out artifacts/perfgate/probes-current.json",
-        args.out_dir.display()
-    );
-    Ok(())
-}
-
-fn write_probe_template_file(path: &Path, content: &str, force: bool) -> anyhow::Result<()> {
-    if path.exists() && !force {
-        anyhow::bail!(
-            "{} already exists; pass --force to overwrite generated probe starter files",
-            path.display()
-        );
-    }
-
-    fs::write(path, content).with_context(|| format!("write probe template {}", path.display()))
-}
-
-#[derive(Debug, Clone, Copy)]
-struct ProbeStarterTemplate {
-    slug: &'static str,
-    title: &'static str,
-    bench: &'static str,
-    scenario: &'static str,
-    workload: &'static str,
-    dominant_probe: &'static str,
-    local_probe: &'static str,
-    support_probe: &'static str,
-    baseline_jsonl: &'static [&'static str],
-    current_jsonl: &'static [&'static str],
-}
-
-fn probe_template(template: ProbeTemplate) -> ProbeStarterTemplate {
-    match template {
-        ProbeTemplate::Parser => ProbeStarterTemplate {
-            slug: "parser",
-            title: "Parser pipeline probes",
-            bench: "parser",
-            scenario: "large_file_parse",
-            workload: "large-file parser throughput",
-            dominant_probe: "parser.batch_loop",
-            local_probe: "parser.tokenize",
-            support_probe: "parser.ast_build",
-            baseline_jsonl: &[
-                r#"{"probe":"parser.tokenize","scope":"local","wall_ms":12.4,"items":10000}"#,
-                r#"{"probe":"parser.ast_build","scope":"local","wall_ms":28.2,"items":10000}"#,
-                r#"{"probe":"parser.batch_loop","scope":"dominant","wall_ms":44.8,"items":10000}"#,
-            ],
-            current_jsonl: &[
-                r#"{"probe":"parser.tokenize","scope":"local","wall_ms":12.8,"items":10000}"#,
-                r#"{"probe":"parser.ast_build","scope":"local","wall_ms":27.5,"items":10000}"#,
-                r#"{"probe":"parser.batch_loop","scope":"dominant","wall_ms":39.6,"items":10000}"#,
-            ],
-        },
-        ProbeTemplate::Batch => ProbeStarterTemplate {
-            slug: "batch",
-            title: "Batch processing probes",
-            bench: "batch",
-            scenario: "batch_transform",
-            workload: "batch transform throughput",
-            dominant_probe: "batch.transform",
-            local_probe: "batch.read_inputs",
-            support_probe: "batch.write_outputs",
-            baseline_jsonl: &[
-                r#"{"probe":"batch.read_inputs","scope":"local","wall_ms":18.0,"items":5000}"#,
-                r#"{"probe":"batch.transform","scope":"dominant","wall_ms":92.0,"items":5000}"#,
-                r#"{"probe":"batch.write_outputs","scope":"local","wall_ms":26.0,"items":5000}"#,
-            ],
-            current_jsonl: &[
-                r#"{"probe":"batch.read_inputs","scope":"local","wall_ms":18.4,"items":5000}"#,
-                r#"{"probe":"batch.transform","scope":"dominant","wall_ms":82.0,"items":5000}"#,
-                r#"{"probe":"batch.write_outputs","scope":"local","wall_ms":25.2,"items":5000}"#,
-            ],
-        },
-        ProbeTemplate::Cli => ProbeStarterTemplate {
-            slug: "cli",
-            title: "CLI workflow probes",
-            bench: "cli-command",
-            scenario: "cli_user_path",
-            workload: "CLI command response path",
-            dominant_probe: "cli.execute",
-            local_probe: "cli.parse_args",
-            support_probe: "cli.render_output",
-            baseline_jsonl: &[
-                r#"{"probe":"cli.parse_args","scope":"local","wall_ms":3.2,"items":1}"#,
-                r#"{"probe":"cli.execute","scope":"dominant","wall_ms":42.0,"items":1}"#,
-                r#"{"probe":"cli.render_output","scope":"local","wall_ms":5.8,"items":1}"#,
-            ],
-            current_jsonl: &[
-                r#"{"probe":"cli.parse_args","scope":"local","wall_ms":3.3,"items":1}"#,
-                r#"{"probe":"cli.execute","scope":"dominant","wall_ms":37.8,"items":1}"#,
-                r#"{"probe":"cli.render_output","scope":"local","wall_ms":5.9,"items":1}"#,
-            ],
-        },
-        ProbeTemplate::Server => ProbeStarterTemplate {
-            slug: "server",
-            title: "Server request probes",
-            bench: "server-request",
-            scenario: "server_local_request",
-            workload: "controlled local server request path",
-            dominant_probe: "server.handle_request",
-            local_probe: "server.decode_request",
-            support_probe: "server.encode_response",
-            baseline_jsonl: &[
-                r#"{"probe":"server.decode_request","scope":"local","wall_ms":4.4,"items":100}"#,
-                r#"{"probe":"server.handle_request","scope":"dominant","wall_ms":31.0,"items":100}"#,
-                r#"{"probe":"server.encode_response","scope":"local","wall_ms":6.5,"items":100}"#,
-            ],
-            current_jsonl: &[
-                r#"{"probe":"server.decode_request","scope":"local","wall_ms":4.5,"items":100}"#,
-                r#"{"probe":"server.handle_request","scope":"dominant","wall_ms":27.0,"items":100}"#,
-                r#"{"probe":"server.encode_response","scope":"local","wall_ms":6.7,"items":100}"#,
-            ],
-        },
-    }
-}
-
-fn render_probe_template_readme(template: &ProbeStarterTemplate, out_dir: &Path) -> String {
-    let out_dir = out_dir.display();
-    format!(
-        r#"# {title}
-
-This starter shows a small probe set for {workload}. Review and edit the probe
-names before using them in durable decision policy.
-
-Generated files:
-
-- `probes-baseline.jsonl`: reviewed baseline probe events
-- `probes-current.jsonl`: current-run probe events for local experimentation
-- `scenario.toml`: scenario snippet to copy into `perfgate.toml`
-- `tradeoff.toml`: tradeoff snippet to copy into `perfgate.toml`
-
-Next:
-
-```bash
-perfgate ingest probes --file {out_dir}/probes-baseline.jsonl --out baselines/probes.json
-perfgate ingest probes --file {out_dir}/probes-current.jsonl --out artifacts/perfgate/probes-current.json
-perfgate probe compare --baseline baselines/probes.json --current artifacts/perfgate/probes-current.json --out artifacts/perfgate/probe-compare.json
-perfgate decision suggest --config perfgate.toml
-```
-
-Do not:
-
-- keep probes that no reviewer can act on;
-- use generated sample values as release proof;
-- turn a temporary debugging span into a durable probe id.
-"#,
-        title = template.title,
-        workload = template.workload,
-        out_dir = out_dir
-    )
-}
-
-fn render_probe_template_scenario(template: &ProbeStarterTemplate) -> String {
-    format!(
-        r#"# Copy into perfgate.toml after reviewing paths and names.
-[[scenario]]
-name = "{scenario}"
-bench = "{bench}"
-weight = 1.0
-probe_baseline = "baselines/probes.json"
-probe_current = "artifacts/perfgate/probes-current.json"
-probe_compare = "artifacts/perfgate/probe-compare.json"
-"#,
-        scenario = template.scenario,
-        bench = template.bench
-    )
-}
-
-fn render_probe_template_tradeoff(template: &ProbeStarterTemplate) -> String {
-    format!(
-        r#"# Copy into perfgate.toml only when reviewers need a bounded tradeoff.
-[[tradeoff]]
-name = "{slug}-dominant-probe-improvement"
-if_failed = "wall_ms"
-downgrade_to = "warn"
-
-[[tradeoff.require]]
-metric = "wall_ms"
-probe = "{dominant_probe}"
-min_improvement_ratio = 1.10
-
-[[tradeoff.allow]]
-metric = "wall_ms"
-probe = "{local_probe}"
-max_regression = 0.03
-
-# Supporting probe to keep visible in reviews: {support_probe}
-"#,
-        slug = template.slug,
-        dominant_probe = template.dominant_probe,
-        local_probe = template.local_probe,
-        support_probe = template.support_probe
-    )
-}
-
 fn execute_probe_compare(args: ProbeCompareArgs) -> anyhow::Result<()> {
     let baseline: ProbeReceipt = read_json(&args.baseline)
         .with_context(|| format!("read baseline probe receipt {}", args.baseline.display()))?;
@@ -3792,503 +3564,6 @@ fn execute_decision_action(
         DecisionAction::Debt(args) => execute_decision_debt(args, server_flags),
         DecisionAction::Export(args) => execute_decision_export(args, server_flags),
         DecisionAction::Prune(args) => execute_decision_prune(args, server_flags),
-    }
-}
-
-fn execute_ledger_action(action: LedgerAction, server_flags: &ServerFlags) -> anyhow::Result<()> {
-    match action {
-        LedgerAction::Doctor(args) => execute_ledger_doctor(args, server_flags),
-    }
-}
-
-fn execute_ledger_doctor(args: LedgerDoctorArgs, server_flags: &ServerFlags) -> anyhow::Result<()> {
-    let config = if args.config.exists() {
-        load_config_file(&args.config)
-            .with_context(|| format!("load ledger doctor config {}", args.config.display()))?
-    } else {
-        ConfigFile::default()
-    };
-    let server_config = server_flags.resolve(&config.baseline_server);
-
-    println!("perfgate ledger doctor");
-    print_ledger_readiness_line(
-        "Local receipts",
-        match ensure_artifact_dir_writable(&args.out_dir) {
-            Ok(()) => format!("ready ({} writable)", args.out_dir.display()),
-            Err(error) => format!(
-                "not ready ({} not writable: {error})",
-                args.out_dir.display()
-            ),
-        },
-    );
-
-    if !server_config.is_configured() {
-        print_ledger_readiness_line("Server URL", "missing");
-        print_ledger_readiness_line("API key", "not configured");
-        print_ledger_readiness_line("Project", "not configured");
-        print_ledger_readiness_line(
-            "Upload mode",
-            "local receipts only; server ledger is optional team history",
-        );
-        print_ledger_readiness_line("History", "not checked; no server URL configured");
-        print_ledger_readiness_line("Export", "not checked; no server URL configured");
-        print_ledger_readiness_line("Prune dry-run", "not checked; no server URL configured");
-        print_ledger_next(&[
-            "You do not need server mode yet.",
-            "Use `perfgate decision bundle` when decision evidence needs to travel.",
-        ]);
-        print_ledger_do_not();
-        return Ok(());
-    }
-
-    let url = server_config
-        .url
-        .as_deref()
-        .expect("checked by is_configured");
-    print_ledger_readiness_line("Server URL", format!("configured ({url})"));
-    print_ledger_readiness_line(
-        "API key",
-        if server_config.api_key.is_some() {
-            "present"
-        } else {
-            "missing; okay only for local unauthenticated server mode"
-        },
-    );
-    let project = server_config.project.as_deref();
-    print_ledger_readiness_line("Project", project.unwrap_or("missing"));
-    print_ledger_readiness_line(
-        "Upload mode",
-        if server_config.fallback_to_local {
-            "advisory (fallback_to_local = true); local receipts remain primary"
-        } else {
-            "server operations are explicit; local receipts remain primary"
-        },
-    );
-
-    if args.offline {
-        print_ledger_readiness_line("Health", "not checked (--offline)");
-        print_ledger_readiness_line("History", "not checked (--offline)");
-        print_ledger_readiness_line("Export", "not checked (--offline)");
-        print_ledger_readiness_line("Prune dry-run", "not checked (--offline)");
-        print_ledger_next(&["Run without `--offline` when you want reachability checks."]);
-        print_ledger_do_not();
-        return Ok(());
-    }
-
-    let client = match ledger_doctor_client(&server_config) {
-        Ok(client) => client,
-        Err(error) => {
-            print_ledger_readiness_line("Health", format!("not checkable: {error:#}"));
-            print_ledger_readiness_line("History", "not checked; client setup failed");
-            print_ledger_readiness_line("Export", "not checked; client setup failed");
-            print_ledger_readiness_line("Prune dry-run", "not checked; client setup failed");
-            print_ledger_next(&["Fix the server URL or use local receipts only."]);
-            print_ledger_do_not();
-            return Ok(());
-        }
-    };
-
-    match with_tokio_runtime(async { client.health_check().await.map_err(anyhow::Error::from) }) {
-        Ok(health) if health.status == "healthy" => {
-            print_ledger_readiness_line("Health", "reachable");
-        }
-        Ok(health) => {
-            print_ledger_readiness_line(
-                "Health",
-                format!("unhealthy response ({})", health.status),
-            );
-        }
-        Err(error) => {
-            print_ledger_readiness_line("Health", format!("unreachable: {error:#}"));
-            print_ledger_readiness_line("History", "not checked; health failed");
-            print_ledger_readiness_line("Export", "not checked; health failed");
-            print_ledger_readiness_line("Prune dry-run", "not checked; health failed");
-            print_ledger_next(&[
-                "Keep local receipts and retry ledger checks after the server is healthy.",
-            ]);
-            print_ledger_do_not();
-            return Ok(());
-        }
-    }
-
-    let Some(project) = project else {
-        print_ledger_readiness_line("History", "not checked; project missing");
-        print_ledger_readiness_line("Export", "not checked; project missing");
-        print_ledger_readiness_line("Prune dry-run", "not checked; project missing");
-        print_ledger_next(&[
-            "Set `--project`, `PERFGATE_PROJECT`, or `[baseline_server].project`.",
-        ]);
-        print_ledger_do_not();
-        return Ok(());
-    };
-
-    let history_result = with_tokio_runtime(async {
-        client
-            .list_decisions(project, &ListDecisionsQuery::new().with_limit(1))
-            .await
-            .map_err(anyhow::Error::from)
-    });
-    match history_result {
-        Ok(response) => {
-            print_ledger_readiness_line(
-                "History",
-                format!(
-                    "reachable ({} decision record(s) visible)",
-                    response.pagination.total
-                ),
-            );
-            print_ledger_readiness_line("Export", "available through decision export");
-        }
-        Err(error) => {
-            print_ledger_readiness_line("History", format!("not reachable: {error:#}"));
-            print_ledger_readiness_line("Export", "not available until history is reachable");
-        }
-    }
-
-    let prune_result = with_tokio_runtime(async {
-        client
-            .prune_decisions(
-                project,
-                &PruneDecisionsRequest {
-                    older_than: chrono::Utc::now(),
-                    dry_run: true,
-                },
-            )
-            .await
-            .map_err(anyhow::Error::from)
-    });
-    match prune_result {
-        Ok(response) => {
-            print_ledger_readiness_line(
-                "Prune dry-run",
-                format!(
-                    "available ({} record(s) matched; dry-run deleted {})",
-                    response.matched, response.deleted
-                ),
-            );
-        }
-        Err(error) => {
-            print_ledger_readiness_line("Prune dry-run", format!("not available: {error:#}"));
-        }
-    }
-
-    print_ledger_next(&[
-        "Use `perfgate decision upload` only after local decision receipts are reviewed.",
-        "Use `perfgate decision history`, `export`, and `prune --dry-run` for team operations.",
-    ]);
-    print_ledger_do_not();
-    Ok(())
-}
-
-fn ledger_doctor_client(server_config: &ResolvedServerConfig) -> anyhow::Result<BaselineClient> {
-    let url = server_config
-        .url
-        .as_deref()
-        .ok_or_else(|| anyhow::anyhow!(BASELINE_SERVER_NOT_CONFIGURED))?;
-    let mut client_config = ClientConfig::new(url)
-        .with_timeout(Duration::from_secs(2))
-        .with_retry(RetryConfig::new().with_max_retries(0));
-    if let Some(api_key) = &server_config.api_key {
-        client_config = client_config.with_api_key(api_key);
-    }
-    BaselineClient::new(client_config)
-        .with_context(|| format!("create ledger doctor client for {url}"))
-}
-
-fn print_ledger_readiness_line(label: &str, value: impl AsRef<str>) {
-    println!("{label}: {}", value.as_ref());
-}
-
-fn print_ledger_next(lines: &[&str]) {
-    println!();
-    println!("Next:");
-    for line in lines {
-        println!("  {line}");
-    }
-}
-
-fn print_ledger_do_not() {
-    println!();
-    println!("Do not:");
-    println!("  make the server ledger part of local correctness.");
-    println!("  rerun benchmarks only to repair optional ledger uploads.");
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DecisionReadiness {
-    RunLocalGateFirst,
-    SimpleGateEnough,
-    PairedModeRecommended,
-    StructuredDecisionCandidate,
-    StructuredDecisionReady,
-    ReadyToBundle,
-}
-
-impl DecisionReadiness {
-    fn status(self) -> &'static str {
-        match self {
-            Self::RunLocalGateFirst => "run_local_gate_first",
-            Self::SimpleGateEnough => "simple_gate_enough",
-            Self::PairedModeRecommended => "paired_mode_recommended",
-            Self::StructuredDecisionCandidate => "structured_decision_candidate",
-            Self::StructuredDecisionReady => "structured_decision_ready",
-            Self::ReadyToBundle => "ready_to_bundle",
-        }
-    }
-
-    fn meaning(self) -> &'static str {
-        match self {
-            Self::RunLocalGateFirst => {
-                "No compare receipts were found yet; run the local gate before making a decision."
-            }
-            Self::SimpleGateEnough => {
-                "The current evidence does not require structured decision ceremony."
-            }
-            Self::PairedModeRecommended => {
-                "The current evidence looks noisy; paired mode is a better next step than decision policy."
-            }
-            Self::StructuredDecisionCandidate => {
-                "Structured decisions may help, but scenario/tradeoff/probe evidence is not ready yet."
-            }
-            Self::StructuredDecisionReady => {
-                "Scenario and tradeoff policy are configured; decision evaluation is useful now."
-            }
-            Self::ReadyToBundle => {
-                "A decision index already exists; the evidence can be exported as a portable bundle."
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-struct DecisionReadinessEvidence {
-    compare_found: usize,
-    compare_missing: usize,
-    has_regression: bool,
-    has_improvement: bool,
-    high_noise: bool,
-    has_probe_config: bool,
-    probe_receipts_found: usize,
-    decision_index_exists: bool,
-}
-
-fn execute_decision_suggest(args: DecisionSuggestArgs) -> anyhow::Result<()> {
-    let config = load_validated_config(&args.config)?;
-    let out_dir = args
-        .out_dir
-        .clone()
-        .unwrap_or_else(|| resolve_configured_out_dir(None, Some(&config)));
-    let evidence = collect_decision_readiness_evidence(&config, &out_dir)?;
-    let readiness = classify_decision_readiness(&config, &evidence);
-    let gaps = decision_readiness_gaps(&config, &evidence);
-
-    println!("perfgate decision suggest");
-    println!();
-    println!("Status: {}", readiness.status());
-    println!("Meaning: {}", readiness.meaning());
-    println!();
-    println!("Evidence:");
-    println!("  benches: {}", config.benches.len());
-    println!("  compare receipts found: {}", evidence.compare_found);
-    println!("  compare receipts missing: {}", evidence.compare_missing);
-    println!("  scenarios: {}", config.scenarios.len());
-    println!("  tradeoff rules: {}", config.tradeoffs.len());
-    println!(
-        "  probe evidence: {}",
-        if evidence.has_probe_config {
-            format!(
-                "configured, {} receipt{}",
-                evidence.probe_receipts_found,
-                plural(evidence.probe_receipts_found)
-            )
-        } else {
-            "not configured".to_string()
-        }
-    );
-    println!(
-        "  decision index: {}",
-        if evidence.decision_index_exists {
-            out_dir.join("decision.index.json").display().to_string()
-        } else {
-            "missing".to_string()
-        }
-    );
-    println!();
-    println!("Structured decisions may help if:");
-    println!("  - one benchmark regressed while another improved");
-    println!("  - reviewers need to accept a bounded tradeoff");
-    println!("  - probe or scenario evidence explains where work moved");
-    if !gaps.is_empty() {
-        println!();
-        println!("Not ready yet:");
-        for gap in &gaps {
-            println!("  - {gap}");
-        }
-    }
-    println!();
-    println!("Next:");
-    for command in decision_readiness_next_commands(readiness, &args.config, &out_dir) {
-        println!("  {command}");
-    }
-    println!("Do not:");
-    println!("  do not make structured decisions mandatory for simple local gates");
-
-    Ok(())
-}
-
-fn collect_decision_readiness_evidence(
-    config: &ConfigFile,
-    out_dir: &Path,
-) -> anyhow::Result<DecisionReadinessEvidence> {
-    let mut compare_found = 0usize;
-    let mut compare_missing = 0usize;
-    let mut has_regression = false;
-    let mut has_improvement = false;
-    let mut high_noise = false;
-
-    for (bench_name, path) in decision_compare_paths(config, out_dir) {
-        if !path.exists() {
-            compare_missing += 1;
-            continue;
-        }
-        compare_found += 1;
-        let compare: CompareReceipt = read_json(&path).with_context(|| {
-            format!("read compare receipt for {bench_name}: {}", path.display())
-        })?;
-        has_regression |= is_regression(compare.verdict.status);
-        has_improvement |= compare.deltas.values().any(|delta| delta.pct < 0.0);
-        high_noise |= compare
-            .deltas
-            .values()
-            .any(|delta| delta.cv.is_some_and(|cv| cv > 0.10));
-    }
-
-    let probe_paths = configured_decision_probe_paths(config);
-    let has_probe_config = !probe_paths.is_empty();
-    let probe_receipts_found = probe_paths.iter().filter(|path| path.exists()).count();
-
-    Ok(DecisionReadinessEvidence {
-        compare_found,
-        compare_missing,
-        has_regression,
-        has_improvement,
-        high_noise,
-        has_probe_config,
-        probe_receipts_found,
-        decision_index_exists: out_dir.join("decision.index.json").exists(),
-    })
-}
-
-fn decision_compare_paths(config: &ConfigFile, out_dir: &Path) -> Vec<(String, PathBuf)> {
-    let single_compare = out_dir.join(COMPARE_RECEIPT_FILE);
-    config
-        .benches
-        .iter()
-        .map(|bench| {
-            let per_bench = out_dir.join(&bench.name).join(COMPARE_RECEIPT_FILE);
-            let path = if config.benches.len() == 1 && single_compare.exists() {
-                single_compare.clone()
-            } else {
-                per_bench
-            };
-            (bench.name.clone(), path)
-        })
-        .collect()
-}
-
-fn configured_decision_probe_paths(config: &ConfigFile) -> Vec<PathBuf> {
-    config
-        .scenarios
-        .iter()
-        .flat_map(|scenario| {
-            [
-                scenario.probe_baseline.as_deref(),
-                scenario.probe_current.as_deref(),
-                scenario.probe_compare.as_deref(),
-            ]
-        })
-        .flatten()
-        .map(PathBuf::from)
-        .collect()
-}
-
-fn classify_decision_readiness(
-    config: &ConfigFile,
-    evidence: &DecisionReadinessEvidence,
-) -> DecisionReadiness {
-    if evidence.decision_index_exists {
-        return DecisionReadiness::ReadyToBundle;
-    }
-    if evidence.compare_found == 0 {
-        return DecisionReadiness::RunLocalGateFirst;
-    }
-    if evidence.high_noise {
-        return DecisionReadiness::PairedModeRecommended;
-    }
-    if !config.scenarios.is_empty() && !config.tradeoffs.is_empty() {
-        return DecisionReadiness::StructuredDecisionReady;
-    }
-    if evidence.has_regression || evidence.has_improvement {
-        return DecisionReadiness::StructuredDecisionCandidate;
-    }
-    DecisionReadiness::SimpleGateEnough
-}
-
-fn decision_readiness_gaps(
-    config: &ConfigFile,
-    evidence: &DecisionReadinessEvidence,
-) -> Vec<&'static str> {
-    let mut gaps = Vec::new();
-    if evidence.compare_found == 0 {
-        gaps.push("no compare receipts found; run `perfgate check` first");
-    }
-    if config.scenarios.is_empty() {
-        gaps.push("no scenario weights configured");
-    }
-    if config.tradeoffs.is_empty() {
-        gaps.push("no tradeoff rules configured");
-    }
-    if !evidence.has_probe_config {
-        gaps.push("no probe evidence configured");
-    } else if evidence.probe_receipts_found == 0 {
-        gaps.push("configured probe evidence was not found on disk");
-    }
-    gaps
-}
-
-fn decision_readiness_next_commands(
-    readiness: DecisionReadiness,
-    config_path: &Path,
-    out_dir: &Path,
-) -> Vec<String> {
-    match readiness {
-        DecisionReadiness::RunLocalGateFirst => vec![format!(
-            "perfgate check --config {} --all --require-baseline",
-            shell_path(config_path)
-        )],
-        DecisionReadiness::SimpleGateEnough => vec![format!(
-            "perfgate check --config {} --all --require-baseline",
-            shell_path(config_path)
-        )],
-        DecisionReadiness::PairedModeRecommended => vec![
-            "perfgate paired --name <bench> --baseline-cmd \"<baseline-cmd>\" --current-cmd \"<current-cmd>\" --repeat 10 --out artifacts/perfgate/<bench>/paired.json".to_string(),
-            format!("perfgate calibrate --config {} --bench <bench>", shell_path(config_path)),
-        ],
-        DecisionReadiness::StructuredDecisionCandidate => vec![
-            "add scenario weights and tradeoff rules for the review question".to_string(),
-            "add probe evidence only if reviewers need to know where work moved".to_string(),
-        ],
-        DecisionReadiness::StructuredDecisionReady => vec![
-            format!("perfgate decision evaluate --config {}", shell_path(config_path)),
-            format!(
-                "perfgate decision bundle --index {}",
-                out_dir.join("decision.index.json").display()
-            ),
-        ],
-        DecisionReadiness::ReadyToBundle => vec![format!(
-            "perfgate decision bundle --index {}",
-            out_dir.join("decision.index.json").display()
-        )],
     }
 }
 
@@ -5542,158 +4817,6 @@ fn execute_cargo_bench(args: CargoBenchArgs) -> anyhow::Result<()> {
     }
 
     Ok(())
-}
-
-fn execute_explain_action(action: ExplainAction) -> anyhow::Result<()> {
-    match action {
-        ExplainAction::Artifacts(args) => execute_explain_artifacts(args),
-    }
-}
-
-fn execute_explain_artifacts(args: ExplainArtifactsArgs) -> anyhow::Result<()> {
-    let mut known = Vec::new();
-    let mut unknown = Vec::new();
-    collect_artifact_files(&args.out_dir, &args.out_dir, &mut known, &mut unknown)?;
-
-    println!("perfgate artifact explanation");
-    println!();
-
-    if !args.out_dir.exists() {
-        println!("Status: no_artifacts");
-        println!(
-            "Meaning: {} does not exist yet; run a check or decision command first.",
-            args.out_dir.display()
-        );
-        println!("Artifacts:");
-        println!("  none");
-        println!("Next:");
-        println!("  perfgate check --config perfgate.toml --all");
-        println!("Do not:");
-        println!("  commit generated artifact directories before reviewing repo policy");
-        return Ok(());
-    }
-
-    if known.is_empty() {
-        println!("Status: no_known_artifacts");
-        println!(
-            "Meaning: {} exists, but no known perfgate receipt files were found.",
-            args.out_dir.display()
-        );
-        println!("Artifacts:");
-        if unknown.is_empty() {
-            println!("  none");
-        } else {
-            for path in &unknown {
-                println!("  {}  unrecognized file", path.display());
-            }
-        }
-        println!("Next:");
-        println!("  perfgate check --config perfgate.toml --all");
-        println!("Do not:");
-        println!("  infer a verdict from unknown files alone");
-        return Ok(());
-    }
-
-    println!("Status: artifacts_found");
-    println!("Meaning: known perfgate receipts or review artifacts are present.");
-    println!("Artifacts:");
-    for (path, role) in &known {
-        println!("  {:<32} {}", path.display(), role);
-    }
-    if !unknown.is_empty() {
-        println!("  unknown:");
-        for path in &unknown {
-            println!("    {}  unrecognized file", path.display());
-        }
-    }
-    println!("Next:");
-    for command in artifact_next_commands(&known) {
-        println!("  {command}");
-    }
-    println!("Do not:");
-    println!("  treat artifacts as durable baselines unless promoted intentionally");
-
-    Ok(())
-}
-
-fn collect_artifact_files(
-    root: &Path,
-    dir: &Path,
-    known: &mut Vec<(PathBuf, &'static str)>,
-    unknown: &mut Vec<PathBuf>,
-) -> anyhow::Result<()> {
-    if !dir.exists() {
-        return Ok(());
-    }
-
-    for entry in fs::read_dir(dir).with_context(|| format!("read {}", dir.display()))? {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_dir() {
-            collect_artifact_files(root, &path, known, unknown)?;
-            continue;
-        }
-
-        let relative = path.strip_prefix(root).unwrap_or(&path).to_path_buf();
-        let Some(name) = path.file_name().and_then(|name| name.to_str()) else {
-            unknown.push(relative);
-            continue;
-        };
-
-        if let Some(role) = known_artifact_role(name) {
-            known.push((relative, role));
-        } else {
-            unknown.push(relative);
-        }
-    }
-
-    known.sort_by(|left, right| left.0.cmp(&right.0));
-    unknown.sort();
-    Ok(())
-}
-
-fn known_artifact_role(file_name: &str) -> Option<&'static str> {
-    match file_name {
-        RUN_RECEIPT_FILE => Some("raw measurement receipt"),
-        COMPARE_RECEIPT_FILE => Some("baseline/current comparison receipt"),
-        "report.json" => Some("machine-readable verdict summary"),
-        "comment.md" => Some("PR-ready human summary"),
-        "repair_context.json" => Some("local reproduction and repair hints"),
-        "decision.md" => Some("human-readable performance decision"),
-        "decision.index.json" => Some("index of decision evidence artifacts"),
-        "decision-bundle.json" => Some("portable decision evidence bundle"),
-        "probe-compare.json" => Some("named probe baseline/current comparison"),
-        "scenario.json" => Some("weighted workload scenario receipt"),
-        "tradeoff.json" => Some("tradeoff policy evaluation receipt"),
-        _ => None,
-    }
-}
-
-fn artifact_next_commands(known: &[(PathBuf, &'static str)]) -> Vec<String> {
-    let has = |name: &str| {
-        known
-            .iter()
-            .any(|(path, _)| path.file_name().and_then(|file| file.to_str()) == Some(name))
-    };
-
-    let mut commands = Vec::new();
-    if has("comment.md") {
-        commands.push("inspect artifacts/perfgate/comment.md or the per-bench comment.md".into());
-    }
-    if has("repair_context.json") {
-        commands.push("inspect repair_context.json for local reproduction and repair hints".into());
-    }
-    if has("decision.index.json") {
-        commands
-            .push("perfgate decision bundle --index artifacts/perfgate/decision.index.json".into());
-    }
-    if has(COMPARE_RECEIPT_FILE) {
-        commands.push("perfgate check --config perfgate.toml --all --require-baseline".into());
-    }
-    if commands.is_empty() {
-        commands.push("perfgate check --config perfgate.toml --all".into());
-    }
-    commands
 }
 
 fn resolve_benchmark_suggestion_profile(
